Added write_noStop and writeStop
As I am not very familiar with arduino (yet) I am really unsure, if the sopy & paste I did makes a lot of sense. Please examine the code carefully.
I removed the part which is now in writeStop from the writeN-function and called it writeN_noStop. I need to get a second test board running to actually test if this works as intended.