Update PCA9634.h
Fixed adressing issues for AllCallAddress. Here the TDS makes a correct claim which was interpreted wrong by me.
In detail: the TDS mentiones the address 0xE0 as AllCall-address. This is correct, however this is an 8 bit address. The LSB (here 0) needs to be ommited when using this library, as it is added during the PCA9634.write() function. Therefore all four addresses (AllCall and Sub1 to Sub3) need to be sortened by the LSB. Then they work flawlessly.