Minor additions for improved communication efficiency (test_86 branch) (#20) (#23)
* Added AllCall and Subcall addresses

Changed missleading comment and added relevant standard addresses of the PCA9634 chip

* Added writeN_noStop

* added writeStop

* Added write_noStop and writeStop

As I am not very familiar with arduino (yet) I am really unsure, if the sopy & paste I did makes a lot of sense. Please examine the code carefully.
I removed the part which is now in writeStop from the writeN-function and called it writeN_noStop. I need to get a second test board running to actually test if this works as intended.

* Small error correction

Code has been tested and seems to work. I am not sure if there are some edge cases which might cause problems...

* Update PCA9634.h

Fixed adressing issues for AllCallAddress. Here the TDS makes a correct claim which was interpreted wrong by me.
In detail: the TDS mentiones the address 0xE0 as AllCall-address. This is correct, however this is an 8 bit address. The LSB (here 0) needs to be ommited when using this library, as it is added during the PCA9634.write() function. Therefore all four addresses (AllCall and Sub1 to Sub3) need to be sortened by the LSB. Then they work flawlessly.

Co-authored-by: Rob Tillaart <rob.tillaart@gmail.com>

Co-authored-by: Luxovious <39154233+Luxovious@users.noreply.github.com>